[frontend] - (Redirection): Move redirection in higher level
